`d_e`: unbind `K^j_v`
The value `K^j_v` inside the hash of `d_e` isn't required for Janus protection,
(see Cypherstack audit), but it *does* make it impossible for the receiver to
efficiently verify Janus protection for normal enotes addressed to a
subaddress of sufficiently high index without the view-incoming key loaded.
Not a huge deal, but it needlessly complicates the scanning code.